# backend/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hair Lux API",
    "version": "1.0.0",
    "description": "API documentation for Hair Lux Booking App"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/hairstyles": {
      "get": {
        "summary": "Get all hairstyles",
        "tags": [
          "Hairstyles"
        ],
        "responses": {
          "200": {
            "description": "List of hairstyles"
          }
        }
      },
      "post": {
        "summary": "Create a new hairstyle",
        "tags": [
          "Hairstyles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Hairstyle created successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
...
    "/api/hairstyles/{id}": {
      "get": {
        "summary": "Get one hairstyle",
        "tags": [
          "Hairstyles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hairstyle details"
          },
          "404": {
            "description": "Hairstyle not found"
          }
        }
      },
      "put": {
        "summary": "Update a hairstyle",
        "tags": [
          "Hairstyles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hairstyle updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a hairstyle",
        "tags": [
          "Hairstyles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hairstyle deleted successfully"
          }
        }
      }
    },
...
  }
}